add prover trait
