+++ Cargo.toml
@@ -18,6 +18,7 @@ path = "src/main.rs"
 strum = "0.24.1"
 strum_macros = "0.24.1"
 halo2_proofs = { git = "https://github.com/junyu0312/halo2", branch = "gpu", default-features = true }
+halo2aggregator-s = { git = "https://github.com/DelphinusLab/halo2aggregator-s.git", tag = "bisect-lookup-0.2.1" }
 halo2ecc-s = { git = "https://github.com/lanbones/halo2ecc-s.git", branch = "bisect-lookup" }
 num-bigint = { version = "0.4", features = ["rand"] }
 poseidon = { git = "https://github.com/lanbones/poseidon" }
